# [Map] Improve web performances by not calling "createInfoWindow" if unnecessary