--- a/src/Map/src/Bridge/Google/assets/src/map_controller.ts
+++ b/src/Map/src/Bridge/Google/assets/src/map_controller.ts
@@ -121,7 +121,13 @@ export default class extends AbstractMapController<
         });
 
         if (infoWindow) {
-            this.createInfoWindow({ definition: infoWindow, marker });
+            if (infoWindow.opened) {
+                this.createInfoWindow({ definition: infoWindow, marker });
+            } else {
+                marker.addListener('click', () => {
+                    this.createInfoWindow({ definition: infoWindow, marker, onMarkerClick: true });
+                });
+            }
         }
 
         return marker;
@@ -130,12 +136,14 @@ export default class extends AbstractMapController<
     protected doCreateInfoWindow({
         definition,
         marker,
+        onMarkerClick,
     }: {
         definition: MarkerDefinition<
             google.maps.marker.AdvancedMarkerElementOptions,
             google.maps.InfoWindowOptions
         >['infoWindow'];
         marker: google.maps.marker.AdvancedMarkerElement;
+        onMarkerClick: boolean;
     }): google.maps.InfoWindow {
         const { headerContent, content, extra, rawOptions = {}, ...otherOptions } = definition;
 
@@ -146,7 +154,7 @@ export default class extends AbstractMapController<
             ...rawOptions,
         });
 
-        if (definition.opened) {
+        if (definition.opened || onMarkerClick) {
             infoWindow.open({
                 map: this.map,
                 shouldFocus: false,
@@ -154,16 +162,11 @@ export default class extends AbstractMapController<
             });
         }
 
-        marker.addListener('click', () => {
+        if (onMarkerClick) {
             if (definition.autoClose) {
                 this.closeInfoWindowsExcept(infoWindow);
             }
-
-            infoWindow.open({
-                map: this.map,
-                anchor: marker,
-            });
-        });
+        }
 
         return infoWindow;
     }