| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | yes <!-- please update src/**/CHANGELOG.md files -->
| Issues        | Fix #... <!-- prefix each issue number with "Fix #", no need to create an issue if none exist, explain below instead -->
| License       | MIT

<!--
Replace this notice by a description of your feature/bugfix.
This will help reviewers and should be a good start for the documentation.

Additionally (see https://symfony.com/releases):
 - Always add tests and ensure they pass.
 - For new features, provide some code snippets to help understand usage.
 - Features and deprecations must be submitted against branch main.
 - Changelog entry should follow https://symfony.com/doc/current/contributing/code/conventions.html#writing-a-changelog-entry
 - Never break backward compatibility (see https://symfony.com/bc).
-->


This afternoon we spoke a bit with @simondaigre, and he told me about InfoWindow performance issues when you have too many of them (and depending your device).

If we have 1.000 InfoWindow, hidden by default and only visible after clicking on a marker, why should be **only** create the InfoWindow at that time?

This PR introduce lazy-creation of InfoWindow, it should improve web-performances on the user's browsers.

I've used the following PHP code:
```php

       $map = (new Map())
            ->center(new Point(0, 0))
            ->zoom(2)
            ->fitBoundsToMarkers()
            ->options(new GoogleOptions(
                mapId: 'XXX',
            ))
        ;

        for ($i = 0; $i < 1_000; $i++) {
            $map->addMarker(
                new Marker(
                    position: new Point(random_int(-90, 90), random_int(-180, 180)),
                    title: 'Marker ' . $i,
                    infoWindow: new InfoWindow(
                        content: 'Info window ' . $i,
                        opened: $i % 499 === 0
                    )
                )
            );
        }
```

## Benchmarks 

I've did some benchmarks for both Google and Leaflet:
- with 1.000 markers, 
- on a Macbook M1 Pro, CPU Throttle x6

### Google

<table>
<tr>
 <td>Before
 <td>After
<tr>
 <td><img width="1800" alt="Capture d’écran 2024-08-18 à 18 28 54" src="https://github.com/user-attachments/assets/816d8bc5-b671-4f6b-a242-28159034c8fc">
 <td><img width="1800" alt="Capture d’écran 2024-08-18 à 18 30 29" src="https://github.com/user-attachments/assets/1f77adae-e0cc-4728-a9f5-4a426d1de6a8">
</table> 

`connect()` takes 936ms instead of ~1,80s, we won ~900ms.

### Leaflet

<table>
<tr>
 <td>Before
 <td>After
<tr>
 <td><img width="1912" alt="Capture d’écran 2024-08-18 à 18 45 10" src="https://github.com/user-attachments/assets/4970c5d8-6e3b-4f3d-8713-f9434a98a661">
 <td><img width="1912" alt="Capture d’écran 2024-08-18 à 18 45 28" src="https://github.com/user-attachments/assets/8a0258cc-811e-4372-817c-43c0a9db765a">
</table> 


`connect()` takes ~1s instead of ~1,20s, we won ~200ms

---

The next steps for performance optimizations are:
- Lazyloading markers and infoWindows, in order to keep a good TTFB 
- Clusters!