@smnandre 2024-08-18
I feel like the current implementation could be improved, but i don't think we could/should avoid exposing this internal code.. and the loop seems a bit odd.

Why not having separate create / open methods ?